+++ src/Plugins/SmartStore.GoogleMerchantCenter/Models/ProfileConfigurationModel.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using SmartStore.Web.Framework;
+
+namespace SmartStore.GoogleMerchantCenter.Models
+{
+	[Serializable]
+	public class ProfileConfigurationModel
+	{
+		public ProfileConfigurationModel()
+		{
+		}
+
+	}
+}